Move initializing logger to logs module
