+++ src/main.rs
@@ -5,7 +5,6 @@ use async_std::stream;
 use async_std::task;
 use comfy_table::presets::{UTF8_FULL, UTF8_HORIZONTAL_BORDERS_ONLY};
 use comfy_table::{ContentArrangement, Table};
-use directories::UserDirs;
 use futures::stream::StreamExt;
 use indicatif::{ProgressBar, ProgressStyle};
 use log::{debug, warn};
@@ -25,34 +24,14 @@ use cli::AppConfig;
 use epub::generate_epubs;
 use extractor::Extractor;
 use http::{download_images, fetch_html};
-use logs::display_summary;
+use logs::{display_summary, init_logger};
 
 fn main() {
     let app_config = cli::cli_init();
 
     if !app_config.urls().is_empty() {
         if app_config.is_debug() {
-            match UserDirs::new() {
-                Some(user_dirs) => {
-                    let home_dir = user_dirs.home_dir();
-                    let paperoni_dir = home_dir.join(".paperoni");
-                    let log_dir = paperoni_dir.join("logs");
-                    if !paperoni_dir.is_dir() || !log_dir.is_dir() {
-                        std::fs::create_dir_all(&log_dir)
-                            .expect("Unable to create paperoni directories on home directory for logging purposes");
-                    }
-                    match flexi_logger::Logger::with_str("paperoni=debug")
-                        .directory(log_dir)
-                        .log_to_file()
-                        .print_message()
-                        .start()
-                    {
-                        Ok(_) => (),
-                        Err(e) => eprintln!("Unable to start logger!\n{}", e),
-                    }
-                }
-                None => eprintln!("Unable to get user directories for logging purposes"),
-            };
+            init_logger();
         }
         download(app_config);
     }
